Move all the enum types into the shared init file for the `ioptron` folder.
